refactor(chat): neutralize low-fanout helper filenames
