+++ turbo/apps/api/src/signals/services/chat-thread-read-state-query.ts
